Create snippet module
Extract the API types for a snippet to a snippet module. Move
the Toggle module over there as well.
