+++ client/components/Editor/Code.tsx
@@ -13,8 +13,9 @@ import {
 } from '../../actions';
 import { selectSelectionStart, selectSelectionEnd } from '../../selectors';
 import { prismSlug } from '../../helpers';
-import { RootAction, Cursor, Toggle } from '../../util';
+import { RootAction, Cursor } from '../../util';
 import { setTheme, togglePlugin } from '../../prism';
+import { Toggle } from '../../snippet';
 import { isSpecialEvent, languageIsEqual, editorOptionsIsEqual } from './util';
 import findOffset from './findOffset';
